🌊 Feat: 파도, 파도타기 삭제 시 시트 내려가게 구현
