+++ PADO/PADO/iOS/Model/User.swift
@@ -9,39 +9,6 @@ import Firebase
 import FirebaseFirestoreSwift
 import Foundation
 
-var userNameID: String {
-    get {
-        UserDefaults.standard.string(forKey: "userNameID") ?? ""
-    }
-    set {
-        UserDefaults.standard.set(newValue, forKey: "userNameID")
-    }
-}
-
-var userToken: String {
-    get {
-        UserDefaults.standard.string(forKey: "userToken") ?? ""
-    }
-    set {
-        UserDefaults.standard.set(newValue, forKey: "userToken")
-    }
-}
-
-var savePhoto: Bool {
-    get {
-        UserDefaults.standard.bool(forKey: "savePhoto")
-    }
-    set {
-        UserDefaults.standard.set(newValue, forKey: "savePhoto")
-    }
-}
-
-var userFollowingIDs: [String] = []
-var acceptAlert = "no"
-var blockingUser: [BlockUser] = []
-var blockedUser: [BlockUser] = []
-var needsDataFetch: Bool = false
-
 struct User: Codable, Identifiable, Hashable {
     @DocumentID var id: String?
     var username: String // 유저 닉네임
